Switch class name from Restaurant to RecipeBook
A previous commit on this branch updated the class Bistro
to be Restaurant because some students were confused about
what a bistro is.  This commit changes the name again to
better match the behavior of the class.  The class does not
represent a restaurant, rather it holds and returns
recipes—like a recipe book.